backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
install==1.3.5
psycopg==3.1.19
psycopg-binary==3.1.19
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.12.2
